automation/requirements.txt
aiohttp==3.9.3
apscheduler==3.10.4
python-dotenv==1.0.1
yt-dlp==2024.3.10
openai-whisper==20231117
gspread==5.12.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openrouter==1.0
python-ffmpeg==2.0.2
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0 